gianfar: Moving inline static function to gianfar.h
Static inline function used by ASF moved to gianfar.h

CQ ID : ENGR00304852
Reviewed-on: http://git.am.freescale.net:8181/10477
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
[Xulin: Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto_RDS_20150206.iso]
